Get rid of Decode_XA. A few minor changes.
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@1235 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ pcsx2/RDebug/deci2.h
@@ -55,7 +55,7 @@ struct DECI2_DBGP_BRK{
 
 extern DECI2_DBGP_BRK	ebrk[32], ibrk[32];
 extern s32 ebrk_count, ibrk_count;
-extern volatile s32 runStatus;
+extern volatile long runStatus;
 extern s32 runCode, runCount;
 
 extern Threading::Semaphore* runEvent;
